requirements.txt
asgiref==3.5.2
Django==4.1.3
django-admin-rangefilter==0.9.0
django-countries==7.4.2
djangorestframework==3.14.0
numpy==1.23.5
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
